Argonaut API - Repackaged model into datatypes, elements, and resources - Made some patient top-level types inner classes of patient
