--- argonaut-api/src/main/java/gov/va/api/health/argonaut/api/Communication.java
@@ -1,27 +1,0 @@
-package gov.va.api.health.argonaut.api;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import java.util.List;
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor
-@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public class Communication implements BackboneElement {
-  @Pattern(regexp = Fhir.ID)
-  String id;
-
-  @Valid List<Extension> modifierExtension;
-  @Valid List<Extension> extension;
-  @NotBlank @Valid CodeableConcept language;
-  Boolean preferred;
-}
