+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/datatypes/Quantity.java
@@ -1,6 +1,8 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.elements.Element;
+import gov.va.api.health.argonaut.api.elements.Extension;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
@@ -16,28 +18,20 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-@Schema(description = "http://hl7.org/fhir/DSTU2/datatypes.html#HumanName")
-public class HumanName implements Element {
+@Schema(description = "http://hl7.org/fhir/DSTU2/datatypes.html#Quantity")
+public class Quantity implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
 
   @Valid List<Extension> extension;
 
-  NameUse use;
-  String text;
-  List<String> family;
-  List<String> given;
-  List<String> prefix;
-  List<String> suffix;
-  @Valid Period period;
+  Double value;
 
-  public enum NameUse {
-    usual,
-    official,
-    temp,
-    nickname,
-    anonymous,
-    old,
-    maiden
-  }
+  @Pattern(regexp = "(<|<=|>=|>)")
+  String comparator;
+
+  String unit;
+
+  @Pattern(regexp = Fhir.CODE)
+  String code;
 }
