+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/datatypes/Range.java
@@ -1,6 +1,8 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.elements.Element;
+import gov.va.api.health.argonaut.api.elements.Extension;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
@@ -16,17 +18,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-@Schema(description = "http://hl7.org/fhir/DSTU2/datatypes.html#SimpleQuantity")
-public class SimpleQuantity implements Element {
+@Schema(description = "http://hl7.org/fhir/DSTU2/datatypes.html#Range")
+public class Range implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
 
   @Valid List<Extension> extension;
-
-  Double value;
-
-  String unit;
-
-  @Pattern(regexp = Fhir.CODE)
-  String code;
+  @Valid SimpleQuantity low;
+  @Valid SimpleQuantity high;
 }
