+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/datatypes/SimpleQuantity.java
@@ -1,6 +1,9 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.elements.Element;
+import gov.va.api.health.argonaut.api.elements.Extension;
+import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
@@ -15,14 +18,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-public class Link implements BackboneElement {
+@Schema(description = "http://hl7.org/fhir/DSTU2/datatypes.html#SimpleQuantity")
+public class SimpleQuantity implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
 
-  @Valid List<Extension> modifierExtension;
   @Valid List<Extension> extension;
-  @Valid Reference other;
+
+  Double value;
+
+  String unit;
 
   @Pattern(regexp = Fhir.CODE)
-  String type;
+  String code;
 }
