+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/elements/Element.java
@@ -1,0 +1,11 @@
+package gov.va.api.health.argonaut.api;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
+
+@Schema(description = "http://hl7.org/fhir/DSTU2/element.html")
+public interface Element {
+  List<Extension> extension();
+
+  String id();
+}
