+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/elements/Meta.java
@@ -1,7 +1,11 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.datatypes.Coding;
+import gov.va.api.health.argonaut.api.elements.Element;
+import gov.va.api.health.argonaut.api.elements.Extension;
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
 import javax.validation.Valid;
 import javax.validation.constraints.Pattern;
 import lombok.AccessLevel;
@@ -15,17 +19,20 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-@Schema(description = "https://www.hl7.org/fhir/resource.html")
-public class SimpleResource implements Resource {
+@Schema(description = "http://hl7.org/fhir/DSTU2/resource.html#meta")
+public class Meta implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
 
-  @Valid Meta meta;
+  @Valid List<Extension> extension;
 
-  @Pattern(regexp = Fhir.URI)
-  @Schema()
-  String implicitRules;
+  @Pattern(regexp = Fhir.ID)
+  String versionId;
 
-  @Pattern(regexp = Fhir.CODE)
-  String language;
+  @Pattern(regexp = Fhir.INSTANT)
+  String lastUpdated;
+
+  List<@Pattern(regexp = Fhir.URI) String> profile;
+  @Valid List<Coding> security;
+  @Valid List<Coding> tag;
 }
