+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/elements/Reference.java
@@ -1,11 +1,11 @@
 package gov.va.api.health.argonaut.api;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.elements.Element;
+import gov.va.api.health.argonaut.api.elements.Extension;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -18,23 +18,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
-@Schema(description = "http://hl7.org/fhir/DSTU2/narrative.html")
-public class Narrative implements Element {
+@Schema(description = "http://hl7.org/fhir/DSTU2/references.html")
+public class Reference implements Element {
   @Pattern(regexp = Fhir.ID)
   String id;
 
   @Valid List<Extension> extension;
-
-  @NotNull NarrativeStatus status;
-
-  @NotBlank
-  @Pattern(regexp = Fhir.XHTML)
-  String div;
-
-  public enum NarrativeStatus {
-    generated,
-    extensions,
-    additional,
-    empty
-  }
+  String reference;
+  String display;
 }
