+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/resources/OperationOutcome.java
@@ -1,0 +1,91 @@
+package gov.va.api.health.argonaut.api;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import gov.va.api.health.argonaut.api.datatypes.CodeableConcept;
+import gov.va.api.health.argonaut.api.elements.BackboneElement;
+import gov.va.api.health.argonaut.api.elements.Extension;
+import gov.va.api.health.argonaut.api.elements.Meta;
+import gov.va.api.health.argonaut.api.elements.Narrative;
+import gov.va.api.health.argonaut.api.resources.DomainResource;
+import io.swagger.v3.oas.annotations.media.Schema;
+import java.util.List;
+import javax.validation.Valid;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@AllArgsConstructor
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+@Schema(description = "https://www.hl7.org/fhir/operationoutcome.html")
+public class OperationOutcome implements DomainResource {
+
+  @Pattern(regexp = Fhir.ID)
+  String id;
+
+  @NotBlank String resourceType;
+  @Valid Meta meta;
+
+  @Pattern(regexp = Fhir.URI)
+  String implicitRules;
+
+  @Pattern(regexp = Fhir.CODE)
+  String language;
+
+  @Valid Narrative text;
+
+  @Valid List<SimpleResource> contained;
+
+  @Valid List<Extension> modifierExtension;
+
+  @Valid List<Extension> extension;
+
+  @NotEmpty @Valid List<Issue> issue;
+
+  @Data
+  @Builder
+  @NoArgsConstructor(access = AccessLevel.PRIVATE)
+  @AllArgsConstructor
+  @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+  @Schema(
+    description = "https://www.hl7.org/fhir/operationoutcome-definitions.html#OperationOutcome.issue"
+  )
+  public static class Issue implements BackboneElement {
+    @Pattern(regexp = Fhir.ID)
+    String id;
+
+    @Valid List<Extension> modifierExtension;
+
+    @Valid List<Extension> extension;
+
+    @NotNull OperationOutcome.Issue.IssueSeverity severity;
+
+    @NotBlank
+    @Pattern(regexp = Fhir.CODE)
+    @Schema(description = "http://hl7.org/fhir/DSTU2/valueset-issue-type.html")
+    String code;
+
+    @Valid CodeableConcept details;
+
+    String diagnostics;
+
+    List<String> location;
+
+    List<String> expression;
+
+    public enum IssueSeverity {
+      fatal,
+      error,
+      warning,
+      information
+    }
+  }
+}
