+++ argonaut-api/src/main/java/gov/va/api/health/argonaut/api/resources/Resource.java
@@ -1,5 +1,6 @@
 package gov.va.api.health.argonaut.api;
 
+import gov.va.api.health.argonaut.api.elements.Meta;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "https://www.hl7.org/fhir/resource.html")
